requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
python-dotenv==1.1.1
Werkzeug==3.1.3
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Werkzeug
gunicorn
python-dotenv
